Fix 6 financial correctness findings from external audit (prompt 2)
- CAT amount rounding: normalize via mojos_to_cat(cat_to_mojos()) in sniper,
  boost_manager, and offer_manager so cached amounts match wallet floor
- Float contamination: int() instead of float() for cat_per_offer in
  wallet_sage.py and wallet_chia.py coin prep
- Session PnL: add since filter to avg_round_trip_secs query in database.py
- Dropped fills: record unknown-detail fills as verification_status='unmatched'
  instead of silently dropping them, so operators can investigate
- Document split_coins_rpc amount_per_coin dead parameter in wallet_sage.py
- Fix test_sniper_coin_ids fake to include mojos_to_cat export

+++ wallet_chia.py
@@ -1463,7 +1463,7 @@ def prepare_coins_for_trading(
         result = split_coins_bulk(
             wallet_id=cat_wallet_id,
             num_coins=cat_needed,
-            coin_size_mojos=float(cat_per_offer),  # For CATs, this is TOKEN amount, not mojos!
+            coin_size_mojos=int(cat_per_offer),  # For CATs, this is TOKEN amount, not mojos!
             fee_mojos=get_effective_transaction_fee_mojos(),
             reserve_multiplier=reserve_multiplier,
             is_cat=True,
